Pull request #34: Cleanup
Merge in TO/sr-1615-it-toads from cleanup to master

Squashed commit of the following:

commit 5a67c6d15061228cabead1177c800edccf095ca5
Author: Jan Zborowski <janzborowski@student.agh.edu.pl>
Date:   Tue Jan 18 11:11:30 2022 +0100

    Address some intelliJ warnings

+++ src/main/java/utils/GameComputations.java
@@ -24,7 +24,7 @@ public class GameComputations {
         Vector2d opposite = translationVector.getSelfOpposite();
         double angle = getConvexAngleWithVerticalAxis(opposite);
         if (opposite.x() > 0) {
-            return (double) (360 - angle);
+            return 360 - angle;
         }
         return angle;
     }
